// frontend/ReceiptTracker/src/app/api/model/receiptOut.ts
/**
 * Receipt Tracker API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserOut } from './userOut';
import { MarketOut } from './marketOut';
import { ReceiptItemOut } from './receiptItemOut';


export interface ReceiptOut { 
    id: number;
    date: string;
    receipt_number: string;
    image_path: string;
    original_filename: string;
    user: UserOut;
    market: MarketOut;
    postal_code: string;
    city: string;
    street_name: string;
    street_number: string;
    items: Array<ReceiptItemOut>;
    total: number;
}