// frontend/ReceiptTracker/src/app/api/model/receiptListOut.ts
/**
 * Receipt Tracker API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {ReceiptOut} from './receiptOut';


export interface ReceiptListOut {
    receipts: Array<ReceiptOut>;
    skip: number;
    limit: number;
    total: number;
    has_next: boolean;
    has_previous: boolean;
}